Move functionality out of celery into management commands
